pyproject.toml
[project]
name = "project-bedrock"
version = "0.1.0"
description = ""
authors = [
    {name = "Pedro Mussi",email = "phpmussi@gmail.com"}
]
readme = "README.md"
requires-python = "~=3.11"
dependencies = [
    "python-dotenv (>=1.1.1,<2.0.0)",
    "requests (<3.0.0)",
    "azure-storage-blob (>=12.26.0,<13.0.0)",
    "pandas (>=2.3.1,<3.0.0)",
    "pyarrow (>=21.0.0,<22.0.0)",
    "boto3 (>=1.40.5,<2.0.0)",
    "databricks-cli (>=0.18.0,<0.19.0)",
    "deltalake (==0.17.0)",
    "clickhouse-connect (>=0.8.18,<0.9.0)",
    "adlfs (>=2023.1.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
